+++ peek_and_pop/example/lib/main.dart
@@ -1,7 +1,15 @@
 //Note: Don't forget to add <key>io.flutter.embedded_views_preview</key><string>YES</string> to your Info.plist. See
 //[webview_flutter](https://pub.flutter-io.cn/packages/webview_flutter) for more info.
 
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// © Cosmos Software | Ali Yigit Bireroglu                                                                                                           /
+// All material used in the making of this code, project, program, application, software et cetera (the "Intellectual Property")                     /
+// belongs completely and solely to Ali Yigit Bireroglu. This includes but is not limited to the source code, the multimedia and                     /
+// other asset files. If you were granted this Intellectual Property for personal use, you are obligated to include this copyright                   /
+// text at all times.                                                                                                                                /
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 //@formatter:off
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';