+++ peek_and_pop/lib/Export.dart
@@ -5,8 +5,8 @@
 // other asset files. If you were granted this Intellectual Property for personal use, you are obligated to include this copyright                   /
 // text at all times.                                                                                                                                /
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
 //@formatter:off
+
 export 'peek_and_pop_controller.dart';
 export 'peek_and_pop_detector.dart';
 export 'peek_and_pop_child.dart';